+++ src/powerquery-language-services/inspection/dereferencedIdentifier/index.ts
@@ -1,0 +1,2 @@
+export * from "./dereferencedIdentifier";
+export * as DereferencedIdentifierUtils from "./dereferencedIdentifierUtils";
